Only import __version__ from mpi4py if has_mpi_support is True to reduce chances that mpi4py will be left in broken state.
Set os.RTLD_DEEPBIND when doing CDLL of liblammps.so in python module
init so that included fake MPI symbols are used in preference to any
other real MPI library that may be loaded by python script (e.g. if it
uses mpi4py but wants to run serial lammps).